feat: resume playback after clipping process is finished
