--- src/ui/shared.scss
@@ -1,27 +1,0 @@
-* {
-    box-sizing: border-box;
-}
-
-html {
-    margin: 0;
-    padding: 0;
-}
-
-body {
-    margin: 0;
-    padding: 16px;
-    font-size: 13px;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-        Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-}
-
-body {
-    color: #1f2023;
-    background: none;
-}
-
-@media (prefers-color-scheme: dark) {
-    body {
-        color: #dadada;
-    }
-}
