+++ src/ui/window/index.js
@@ -1,5 +1,3 @@
-import "../shared.scss";
-
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./app.jsx";
@@ -7,5 +5,5 @@ import App from "./app.jsx";
 import '@fontsource/inter';
 
 document.addEventListener("DOMContentLoaded", () => {
-  ReactDOM.render(<App>Hello, world!</App>, document.getElementById("root"));
+  ReactDOM.render(<App />, document.getElementById("root"));
 });
